+++ src/components/helpers/random.ts
@@ -1,4 +1,6 @@
-const counter = 0;
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 function dec2hex (dec: number) {
     return dec < 10